ai-services/fastapi-ai/pyproject.toml
[project]
name = "fastapi-ai-services"
version = "0.1.0"
description = "FastAPI services for AI-01 risk prediction and AI-02 sentiment analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "python-dotenv==1.0.1",
    "scikit-learn>=1.5.0",
]

[dependency-groups]
dev = [
    "pytest==8.3.3",
]
train = [
    "torch>=2.3.0",
    "transformers>=4.44.0",
    "datasets>=2.20.0",
    "accelerate>=0.33.0",
    "pandas>=2.2.0",
    "pyyaml>=6.0.2",
]
